Renamed instances of Basic Rule to Direct Rule
